Created EmptySet, UniversalSet, and implemented most functionality in the AbstractSet libraries. Next, need to make AbsoluteSet and RelativeSet interoperable. Changed AbsoluteSet to never grow its universe implicitly, and if we ever need that functionality, it will be the responsibility of FiniteUniverseSet to do the universe-merging stuff.
Successfully interfaced AbsoluteSet with ElementDef and ElementUse. The
default ElementUse#allowed_values is the UniversalSet which is an infinite
set that #includes? all possible values. For the ID element types with
internal code lists, this is overridden to produce an AbsoluteSet with a
universe set consisting of all codes in the list.

Moved helper functions to GuideBuilder module.