Started filling in parts of the AbstractClass.instructions methods that are called when pushing new states onto the stack. Discovered some complications when needing to push several states for a single transition (eg reading "ST" pushes a transaction set and the header table, because "ST" actually belongs to the table).
Implemented most of the functionality that lets the StateMachine execute
individual Instruction values.

Implemented the wiring between the TokenReader and the AbstractStates
which is quite a bit cleaner than before, and should also support the
functionality needed for programmatic generation. When we read "ISA"
it allows us to derive the SegmentDefs for the specific version, as
well as the version-specific delimiters. When we read "GS" we can then
derive the rest of the SegmentDefs for that X12 version (eg 5010). The
TokenReader uses this information because it can't know when elements
are composite or repeating from the input alone.

Documented most of the Builder::Instruction class