Converted RDoc-formatted comments to Markdown
Added usage to RepeatedElementVal, which is needed because the
usage cannot be derived unless the element isn't empty.

Improved documentation in various parts of the code base. Mostly
finished 'Generating.md', which describes how to generate X12 using
Builder::BuilderDsl.

Created a Reader::Position class that will be used to track the source
of each token (and the call stack when generating X12). Currently the
instance of Reader::Input is being used, which isn't ideal.