Created a more convenient constructor Reader.build to construct a Reader::Input value and a Reader::StreamReader around it
Fixed most of the problems with Navigation methods, which were due
to inconsistent pointers from states to values. The number of edits
needed to bring the state tree in line with the value tree is a
potential problem, but there appears to already be plenty of room
for optimization at this point.