Split the StateMachine into three mixins: Tokenization, Generation, and Navigation. Most of the code in Tokenization was removed from the BuilderDsl, which now adds very little -- need to revisit it soon.
Enhanced the InstructionTable#matches methods and the supporting
ConstraintTable#matches methods to (optionally) throw an exception
when the constructed SegmentTok has elements that aren't in their
declared #allowed_values set. This is a temporary implementation,
which isn't quite right because another state's InstructionTable
might actually allow that element value.

Commented and cleaned up several parts of InstructionTable#drop

Navigation#find now works with ConstraintTable to limit the number of
possible states that need to be searched. This allows searching for
segments based on qualifiers:

  # Find the first HL segment of any type
  parser.first.
    flatmap{|x| x.find(:GS) }.
    flatmap{|x| x.find(:ST) }.
    flatmap{|x| x.find(:HL, nil, nil, "22") }

  # Find the first subscriber HL segment
  parser.first.
    flatmap{|x| x.find(:GS) }.
    flatmap{|x| x.find(:ST) }.
    flatmap{|x| x.find(:HL, nil, nil, "22") }

  # Throws an exception, "30" is never an allowed value
  parser.first.
    flatmap{|x| x.find(:GS) }.
    flatmap{|x| x.find(:ST) }.
    flatmap{|x| x.find(:HL, nil, nil, "30") }

But it's unclear what should be done with given elements that aren't
characteristic to a certain segment. For instance, x.find(:HL, "1") doesn't
narrow the search because there are no constraints on the first element.

Minor enhancements to ThreadLocalVar. This looked like a good fit to
prevent threading issues with InstructionTables when a StateMachine
is shared among threads, but I think it's actually safe to let the
threads stomp on one another since race conflicts will all result in
a consistent object state.