Improved comments and documentation of ConstraintTable, and fixed an error message
Implemented Navigation#segment and Navigation#element to read parts
of the parse tree that are subordinate to the current position.

Fixed Navigation#first and Navigation#last to return Either.failure
when there are no segments in the parse tree (eg, it's empty)

Renamed RootConfig to Config. Changed some ParseError exceptions to
ArgumentErrors, etc.

Fixed code values for the TA1 note element... copy pasta gone wrong.

Started implementing automated generation of FA999 and TA1
acknowledgements... it's not going to be pretty.