Enhanced FiveOhOneEd and FiftyTenEd to find invalid envelopes using Navigation#find!, and for example, report "unrecognized version '004010'"
Fixed a bug that tried to combine a time with an invalid date to
check if the date/time was in the future in ISA and GS segment edits

Changed Navigation#element to return an Either.failure when the
current segment isn't valid. The invalid segment's element tokens are
still accessable by other means, but are generally useless.

Fixed Navigation#first and Navigation#last methods so they don't
return Either.failure when the first or last segment is invalid