Added an optional boolean parameter to #to_x12 that, when true, will truncate the element to the defined maximum length. This won't prevent BuilderDsl from raising an exception that a generated element is too long (because the implementation of #too_long? hasn't changed) but it can prevent certain problems
Added a #map method to several SimpleElementVal subclasses to
avoid having to define, for example, StringVal#downcase, or hundreds
of other methods such that they return another StringVal. This would
go hand-in-hand with rewriting elements in the parse tree, if such a
thing were implemented.

Started implementing specs for each SimpleElementVal subclass