Fix #117 JRuby incompatibilities
It seems like JRuby doesn't correctly implement refinements see
(jruby/jruby#4330) so the only way around this was to revert to
monkey-patching the base classes. New file lib/ruby/jruby_hack.rb
is copy-pasted code from each lib/ruby file, except instead of
'refine X do { ... }' it's changed to 'class X { ... }'.

There was also an incompatibility with Nn.to_d - Nn, which fails
with "Nn cannot be coerced to BigDecimal"; this might be an error
in the BigDecimal class on JRuby? For now, those specs are marked
pending on JRuby.

Redcarpet is a C-extension for MRI, so I disabled it for JRuby,
but documentation can be accessed online (just not built on JRuby)