Update time_val.rb
If the object is 6 characters long, decimal is set to an empty string on line 293 (object.to_s.slice(6..-1) is an empty string). Since decimal is set, the if statement is true and the program steps into the block. In the block we now have "0.".to_d, which results in an ArgumentError (invalid value for BigDecimal(): "0.").

This change fixes this issue, as if decimal is an empty string, it will be changed to 0 before to_d is called on it.