Address questions in #172 regarding tables: if a table begins with a repeatable loop, then the table is also considered repeatable; however, the parser will prefer to reuse the current table, if possible, when instantiating new instances of that loop. Otherwise, when the table starts with a segment (optional or mandatory), or a non-repeatable loop, then the table cannot be repeated.
Fix an issue (previously irrelevant) so tables can begin with not only
the first required SegmentUse or LoopDef declared, but all segments or
loops that start at the same position. Now tables that begin with a
qualified loop (eg, 2000A, 2000B, 2000C) work with the above changes.

Make a change so the parser, when inserting a segment, will always
figure out a single Instruction to execute, but when looking for an
existing segment it will execute an Instruction for each possibility
(eg, when a loop's start segment could've been placed in either a new
table or the current one).

Restore 005010X212 HR276 and 005010X214 HN277 to their ideal state,
now that TableDef.detail infers whether the table is repeatable or not,
and now that the parser accomodates both cases correctly.

Write specs for Navigation#find that enumerate most possible cases with
a minimal grammar defined for each case